# frontend/src/data/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store CRUD API",
    "version": "1.0.0",
    "description": "A comprehensive REST API for managing store inventory, dynamic tables, and e-commerce functionality.",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/tables": {
      "get": {
        "tags": ["Tables"],
        "summary": "List Tables",
        "description": "Retrieve user's dynamic tables with metadata",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of tables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Table"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tables"],
        "summary": "Create Table",
        "description": "Create a new dynamic table with custom schema",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateTableRequest"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Table created successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Table"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token authentication. Format: `Bearer <token>`"
      }
    },
    "schemas": {
...
      "Table": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique table identifier"
          },
          "name": {
            "type": "string",
            "description": "Table name"
          },
          "description": {
            "type": "string",
            "description": "Table description"
          },
          "for_sale": {
            "type": "boolean",
            "description": "Whether this is an e-commerce table"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTableRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Table name"
          },
          "description": {
            "type": "string",
            "description": "Table description"
          },
          "for_sale": {
            "type": "boolean",
            "description": "Whether this is an e-commerce table",
            "default": false
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}